Moved circle and arrows to sep files
